requirements.txt
discord.py
python-dotenv
Flask
